Remove reference to TreeSet from SortedSet$.
Due to signature issues in RedBlack, TreeSet is evicted by GWT, and if SortedSet
gets evicted, then so does scala.Enumeration and most of the rest of the
library.

Snipping the dependency of SortedSet$ on TreeSet stops the eviction cascade at a
place where most of the library remains usable.